% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice_qtl.R
\docType{data}
\name{rice_qtl}
\alias{rice_qtl}
\title{Real experimental data}
\format{A data frame 326 rows (individual) and 38 variables (SNPs)}
\source{
{Uruguayan Rice Breeding GWAS (URiB)}
}
\usage{
rice_qtl
}
\description{
The data is a QTL for rice Grain Quality
}
\keyword{datasets}

