% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_qtl.R
\docType{data}
\name{sim_qtl}
\alias{sim_qtl}
\title{simple QTL simulated}
\format{A data frame 5 rows (individuals) and 7 variables (snps)}
\source{
{simulated data}
}
\usage{
sim_qtl
}
\description{
A dataset containing the marcadores
}
\keyword{datasets}

